pyproject.toml
[tool.poetry]
name = "chirp"
version = "0.1.0"
description = "A bioacoustics research project."
authors = ["Chirp Team <chirp-bio@google.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
# These dependencies are part of the implicit 'main' group, meaning
# they will be installed when using `pip install`.
# Other dependency groups will not be installed by pip, however.
# See: https://python-poetry.org/docs/managing-dependencies/#dependency-groups
python = ">=3.10,<3.12"
absl-py = "^1.4.0"
apache-beam = {version = "^2.50.0", extras = ["gcp"]}
clu = "^0.0.9"
flax = "^0.8.1"
jax = "^0.4.16"
ml-collections = "^0.1.1"
optax = ">=0.1.7"
#tensorflow = "^2.15"
pandas = {version = "^2.1.1", extras=["gcp"]}
etils = {version = "^1.5.0", extras = ["epath"]}
SPARQLWrapper = "^2.0.0"
tensorflow-hub = "^0.16.0"
imageio = "^2.5.0"
matplotlib = "^3.6.1"
notebook = "^7.0.4"
chex = "^0.1.7"
numba = "^0.57"
librosa = "^0.10.1"
numpy = "^1.24.0"
tqdm = "^4.6"
ipywidgets = "^8.1"
scann = "^1.3.2"

[tool.poetry.group.jaxtrain]
optional = true

[tool.poetry.group.jaxtrain.dependencies]
# TF-io is used for sample rate conversion only in data ingestion pipelines.
# Note that tensorflow and tensorflow-io versions must be synchronized.
# See: https://github.com/tensorflow/io#tensorflow-version-compatibility
#tensorflow = "^2.16"
tensorflow-io = "^0.37"
tensorflow-datasets = {version = "^4.9.3", extras = ["dev"]}
# AudioMAE / ViT.
scenic = {git = "https://github.com/google-research/scenic.git"}
# Used for crawling datasources.
ratelimiter = "^1.2.0.post0"
# Quantization-aware training in Jax.
aqtp = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
